+++ src/main/java/commuml/Application.java
@@ -1,4 +1,4 @@
-package hello;
+package commuml;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
