+++ src/main/java/commuml/DrawingController.java
@@ -1,4 +1,4 @@
-package hello;
+package commuml;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
