+++ src/main/java/commuml/MainController.java
@@ -1,9 +1,8 @@
-package hello;
+package commuml;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
 public class MainController {
